A3/Q1/requirements.txt
einops==0.6.1
matplotlib==3.7.2
numpy==1.23.5
opencv_python==4.8.0.74
PyYAML==6.0.1
torch==1.11.0
torchvision==0.12.0
tqdm==4.65.0
pillow>=9.0.0
